update pyperdrive to match new hyperdrive-rs functions (#7)
This PR updates the pyperdrive wrappers to match the latest hyperdrive-rs functions. I removed a test because the new `get_max_short` workflow includes a private `max_short` function, the output of which is required to conduct the test. This functionality should be tested in the underlying rust code, although we should still include a suite of tests to guarantee that errors are propagated into Python correctly.